requirements.txt
asgiref==3.7.2
Django==4.2.6
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-cors-headers==4.3.0
django-flat-responsive==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-slugify==8.0.1
pytz==2023.3.post1
sqlparse==0.4.4
text-unidecode==1.3
